requirements.txt
Django
pillow
django-cleanup
django-allauth
django-htmx
channels-redis
daphne
shortuuid
psycopg2
python-dotenv
whitenoise